token logic moved in createTokenWithAttachCookies
