+++ controllers/auth/signin.js
@@ -1,7 +1,5 @@
-const randtoken = require('rand-token');
-
-const { Users, Tokens } = require('../../db/models');
-const { createToken } = require('./createToken');
+const { Users } = require('../../db/models');
+const { createTokenWithAttachCookies } = require('./createTokenWithAttachCookies');
 
 async function signin(req, res) {
   const { body: { email, password } } = req;
@@ -15,20 +13,9 @@ async function signin(req, res) {
         message: 'Authentication failed. User not found.',
       });
     }
-    const isCompare = user.comparePassword();
+    const isCompare = user.comparePassword(password);
     if (isCompare) {
-      const token = createToken(user.id);
-      const refresh = randtoken.uid(255);
-      await Tokens.create({
-        userId: user.id,
-        token: refresh,
-      });
-
-      res.cookie('refreshToken', refresh, {
-        secure: false,
-        maxAge: 30 * 24 * 60 * 60 * 1000,
-        httpOnly: true,
-      });
+      const token = await createTokenWithAttachCookies(user.id, res);
       return res.json({ success: true, token, user });
     }
     return res.status(401).send({ success: false, msg: 'Authentication failed. Wrong password.' });
